Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-2
> 
> this patch and commit message do not seem to be the same thing.

The Jira is about moving Micro from Florine to Latest Master. This ideally requires version changes to the POM files.
But from Florine to current master, there are lot of changes to MDSAL and controller packages. Some of the deprecated code in controller is no longer available in the current master and we have to replace them with the MDSAL APIs.
So this patch also has those API migrations.

Patch-set: 1